Initial work on tweaking the gem to support multiple neo4j instances.
