+++ lib/neoid/database_cleaner.rb
@@ -1,7 +1,9 @@
 module Neoid
   class NeoDatabaseCleaner
-    def self.clean_db(start_node = Neoid.db.get_root)
-      Neoid.db.execute_script <<-GREMLIN
+    def self.clean_db(instance, start_node = nil)
+      puts instance.inspect
+      start_node ||= instance.db.get_root
+      instance.db.execute_script <<-GREMLIN
         g.V.toList().each { if (it.id != 0) g.removeVertex(it) }
         g.indices.each { g.dropIndex(it.indexName); }
       GREMLIN
