+++ lib/neoid/railtie.rb
@@ -4,7 +4,7 @@ module Neoid
   class Railtie < Rails::Railtie
     initializer "neoid.configure_rails_initialization" do
       config.after_initialize do
-        Neoid.initialize_all
+        Neoid::Connections.initialize_all
       end
     end
